// src/app/core/api/generated/model/ingredientCategoryResponse.ts
/**
 * Precify API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Resposta de uma categoria de ingrediente.
 */
export interface IngredientCategoryResponse { 
    /**
     * Identificador da categoria.
     */
    id?: string;
    /**
     * Nome da categoria.
     */
    name?: string | null;
    /**
     * Indica se a categoria permanece ativa.
     */
    isActive?: boolean;
    /**
     * Data de criação da categoria.
     */
    createdAt?: string;
    /**
     * Data da última atualização da categoria.
     */
    updatedAt?: string;
}